% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineQuality.R
\name{wineQuality}
\alias{wineQuality}
\title{The Wine Quality Dataset Generator Function}
\usage{
wineQuality(nObservations = 100, scoreCutOff = 25)
}
\arguments{
\item{nObservations}{The total number of observations requested. Defaults to 100}

\item{scoreCutOff}{The cutoff score set to determine whether a wine is good. Defaults to 25}
}
\description{
This function generates a mock wine quality survey. The function generates data for a total of 9 different countries and returns the total number of wines surveyed and the number of wines that were labelled as good for each country.
}
\examples{
wineQuality(scoreCutOff = 50)
}
\keyword{FDVMidterm}
\keyword{mockdata}
\keyword{wineQuality}
