% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineQualityModel.R
\name{wineQualityModel}
\alias{wineQualityModel}
\title{The Wine Quality Model Results}
\usage{
wineQualityModel(nObservations = 1000, seed = 1)
}
\arguments{
\item{nObservations}{The number of requested observations to generate. Defaults to 1000}

\item{seed}{Sets the seed of the random number generator (you will be instructed on what value to use here in your midterm). Defaults to 1}
}
\description{
This function builds a logistic regression from mock wine quality data. We aim, with the data to  build a model that can accurately predict wine quality according to its variety, fixed acidity, sulfates, and alcohol level. This function builds a data set for you, trains a classifier (logistic regression) and returns the predictions of the model in addition to the test data set. The relevant columns to use here are GoodWine (the outcome) and prediction (the model prediction).
}
\examples{
wineQualityModel(nObservation = 1000, seed = 1)
}
\keyword{logisticregression}
\keyword{midterm}
\keyword{modelperformance}
